requirements.txt
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
auto_gptq==0.7.1
av==13.1.0
azure-ai-documentintelligence==1.0.0b4
certifi==2024.8.30
charset-normalizer==3.4.0
coloredlogs==15.0.1
colpali_engine==0.3.8
datasets==3.1.0
dill==0.3.8
distro==1.9.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
faiss-cpu==1.9.0
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.2.0
gekko==1.2.1
gliner
GPUtil==1.4.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
Jinja2==3.1.3
jiter==0.7.0
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
mteb==1.6.35
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.58.1
optimum==1.23.3
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
peft
pillow==10.2.0
polars==1.12.0
propcache==0.2.0
psutil==6.1.0
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytrec-eval==0.5
pytrec-eval-terrier==0.5.6
pytz==2024.2
PyYAML==6.0.2
qwen-vl-utils==0.0.8
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rouge==1.0.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers<3.0.0,>=2.2.2
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.0
transformers
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xxhash==3.5.0
yarl==1.17.1
nest_asyncio
databricks-sdk